# server/docs/openapi.yml
openapi: 3.0.0
info:
  title: Vector Graphic Editor API
  description: API for managing canvas in vector graphic editor
  version: 1.0.0

servers:
  - url: http://localhost:9090/api
    description: Development server

paths:
  /canvas:
    get:
      summary: Get all canvases metadata
      operationId: getCanvasesMeta
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 200
                data:
                  - id: "102303"
                    createdAt: "2026-02-23T10:30:00"
                    updatedAt: "2026-02-23T13:45:00"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 500
                message: "Internal server error: Database connection failed"

    post:
      summary: Create a new canvas
      operationId: createCanvas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CanvasDto'
            example:
              content: "{}"
      responses:
        '201':
          description: Canvas created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 201
                data:
                  id: "102303"
                  content: "{}"
                  createdAt: "2026-02-23T10:30:00"
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 400
                message: "Not all required fields are represented"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 500
                message: "Internal server error: Database connection failed"

  /canvas/{id}:
    get:
      summary: Get canvas by ID
      operationId: getCanvasById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "102303"
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 200
                data:
                  id: "102303"
                  content: "{}"
                  createdAt: "2026-02-23T10:30:00"
                  updatedAt: "2026-02-23T13:32:00"
        '404':
          description: Canvas not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 404
                message: "Canvas not exist"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 500
                message: "Internal server error: Database connection failed"

    put:
      summary: Update an existing canvas
      operationId: updateCanvas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "102303"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CanvasDto'
            example:
              content: "{\"elements\":[{\"type\":\"rectangle\"}]}"
      responses:
        '200':
          description: Canvas updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 200
                data:
                  id: "102303"
                  content: "{\"elements\":[{\"type\":\"rectangle\"}]}"
                  createdAt: "2026-02-23T10:30:00"
                  updatedAt: "2026-02-23T10:45:00"
        '404':
          description: Canvas not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 404
                message: "Canvas not exist"
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 400
                message: "Invalid content format"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 500
                message: "Internal server error: Database connection failed"

    delete:
      summary: Delete a canvas
      operationId: deleteCanvas
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          example: "102304"
      responses:
        '204':
          description: Canvas deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 204
                message: Deleted successfully
        '404':
          description: Canvas not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 404
                message: "Canvas not exist"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response'
              example:
                code: 500
                message: "Internal server error: Database connection failed"

components:
  schemas:
    CanvasDto:
      type: object
      description: Canvas data transfer object for create/update operations
      properties:
        id:
          type: string
          description: Unique canvas identifier
          example: "102303"
        content:
          type: string
          description: Canvas content in JSON format
          example: "{}"
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2026-02-23T10:30:00"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2026-02-23T10:45:00"
      required:
        - content

    CanvasMeta:
      type: object
      description: Canvas metadata without content
      properties:
        id:
          type: string
          description: Unique canvas identifier
          example: "102303"
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2026-02-23T10:30:00"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2026-02-23T10:45:00"
      required:
        - id
        - createdAt
        - updatedAt

    Response:
      type: object
      description: Standard API response wrapper
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 200
        message:
          type: string
          nullable: true
          description: Error message or null for successful operations
          example: null
        data:
          oneOf:
            - type: array
              items:
                $ref: '#/components/schemas/CanvasMeta'
            - $ref: '#/components/schemas/CanvasDto'
            - $ref: '#/components/schemas/CanvasMeta'
          nullable: true
          description: Response payload
      required:
        - code
      example:
        code: 200
        data:
          - id: "102303"
            createdAt: "2026-02-23T10:30:00"
            updatedAt: "2026-02-23T10:345:00"